feat: 당첨 번호 입력 예외 테스트 추가 및 검증 구현
