+++ README.md
@@ -7,18 +7,14 @@
 - [x] 당첨 번호를 입력 받는다. 번호는 쉼표(,)를 기준으로 구분한다.
     - 예외
         - [x] 입력된 번호의 숫자 범위가 1~45 밖인 경우
-        - [] 6자리 숫자가 아닌 경우
+        - [x] 6자리 숫자가 아닌 경우
+        - [x] 중복되는 숫자가 있을 경우
 - [] 보너스 번호를 입력 받는다.
     - 예외
         - [x] 입력된 번호의 숫자 범위가 1~45 밖인 경우
 
 ### 로또 발행
-- [] 1개의 로또를 발행할 때 중복되지 않는 6개의 숫자를 뽑는다.
-  - 예외
-    - [] 로또 번호의 숫자 범위가 1~45 밖인 경우
-
-### 당첨 번호 뽑기
-- [] 당첨 번호 추첨 시 중복되지 않는 숫자 6개와 보너스 번호 1개를 뽑는다.
+- [] 구입 금액에 따라 로또를 발행한다.
 
 ### 당첨 계산
 - [] 로또 번호와 당첨 번호를 비교해 당첨 내역과 수익률을 출력한다.
@@ -36,7 +32,7 @@
 ### 출력
 - [] 발행한 로또 수량 및 번호를 출력한다. 로또 번호는 오름차순으로 정렬하여 보여준다.
 - [] 당첨 내역을 출력한다.
-- [] 수익률은 소수점 둘째 자리에서 반올림한다. (ex. 100.0%, 51.5%, 1,000,000.0%)
+- [] 수익률은 소수점 둘째 자리에서 반올림해 출력한다. (ex. 100.0%, 51.5%, 1,000,000.0%)
 - [x] 예외 상황 시 에러 문구를 출력해야 한다. 단, 에러 문구는 "[ERROR]"로 시작해야 한다.
 - [] Exception이 아닌 IllegalArgumentException, IllegalStateException 등과 같은 명확한 유형을 처리한다.
